+++ src/main/java/com/kernel360/boogle/book/db/BookEntity.java
@@ -1,4 +1,4 @@
-package com.kernel360.boogle.entity;
+package com.kernel360.boogle.book.db;
 
 
 import lombok.AllArgsConstructor;
